Add support for default methods on builder interfaces
Two different approaches are required for Java 8 and Java 9+.

The technique for Java 9+ involves an illegal reflective operation which
will have the runtime complain when the default method is called. This
may result in broken behavior in a future Java release although it should
work for now with Java 9 and Java 10.

Ref: https://stackoverflow.com/questions/37812393/how-to-explicitly-invoke-default-method-from-a-dynamic-proxy

Fixes #11